LayoutTests/fast/js/sputnik/Conformance/08_Types/8.3_The_Boolean_Type/S8.3_A2.1-expected.txt

S8.3_A2.1

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
